LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-range-ignored-expected.txt


FAIL Animation API call rangeStart overrides animation-range-start assert_approx_equals: values do not match for "animation's start time after first set of range updates" expected 16.666666666666668 +/- 0.125 but got 0
FAIL Animation API call rangeEnd overrides animation-range-end assert_approx_equals: values do not match for "iteration duration after first style change" expected 50 +/- 0.125 but got 16.666667938232422
